raidboss: call to get towers after poison n pop expires
During beat 3, players get a "Poison 'n' Pop" debuff. When this debuff
expires on a player, a huge aoe will trigger. 4 players get a long
version of this debuff, and 4 players get a short version of this
debuff. When each set expires, the 4 players who didn't get hit by an
aoe need to quickly run from the middle to grab a tower.

Use a collec trigger to capture when players lose the debuff effect.
Then in the callout trigger, inform anyone else not in the collected
list to get their tower.

Ideally, we could add a trigger to inform when the players should go out
with their tower. This is tricky, as it either needs to be called after
a stage combo, or somehow timer based. Both of these are tricky to
implement, so have been left as a futer improvement.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>