personal: add some callouts for mount rokkon
Add callouts for the follow up mechanic from stormcloud summons. This is
my best attempt at a simple callout. There isn't much time to get in
position, but its not quite clear if we can do better.

Perhaps the data from the smokeater could provide a better indication of
where the safe spots are, but its not clear if we actually get good data
from this.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>