raidboss: introduce helper function for stored cleaves
In m4n, we need to take a list of stored cleave directions and convert
them into cardinal safe spots. Similar code is used in two places. A
future change is going to refactor this logic.

Make StoredCleave an explicit type, and introduce a getCleaveDirs helper
function which takes the actors and stored cleave data, and returns an
array of DirectionOutputCardinal names.

This avoids duplicating the somewhat complex logic of looking up actor
data and finding the facing. It also makes it easier to extend this
logic in the near future.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>