raidboss: add support for DN strategy during Witch Hunt in m4s
During Witch Hunt, players need to bait aoes so that no player takes the
Witch Hunt AoE twice. Currently, we callout either "Baits near (Party
Far)" or "Baits Far (Party Near)".

Users can customize these if they use a strategy which has fixed
ordering (say, Tanks, Healers, Melee, Ranged). However, the standard NA
strat is the "DN" strat which uses a flexible ordering for update, where
Tanks and Melees prioritize the close/in spots. This cannot be
implemented as-is just with custom strings.

Add a new strategy option which implements the DN priority bate
strategy. I tried to make a shared function it ended up being worse than
just duplicating the local lambda function used here.

In addition to modifying the alert callout for when to be in/out and
close/far, also modify the combined starting callout. Instead of calling
"close/far" for the baits, call the order of who baits, i.e.:

  "In (Tanks) => Out (Healers) => In (Melee) => Out (Ranged)"

Note that we don't need to call "tanks close/far" because this strategy
always has tanks and melee baiting close while range and healers bait
far.

For DPS players, the strategy always calls out with the DPS baits first.
This is to account for the fact that sometimes DPS players of the
opposite role will fill a spot (i.e. a ranged player taking a melee spot
if the party has 3 ranged players).

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>