raidboss: add support for DN strategy during Witch Hunt in m4s
During Witch Hunt, players need to bait aoes so that no player takes the
Witch Hunt AoE twice. Currently, we callout either "Baits near (Party
Far)" or "Baits Far (Party Near)".

Users can customize these if they use a strategy which has fixed
ordering (say, Tanks, Healers, Melee, Ranged). However, the standard NA
strat is the "DN" strat which uses a flexible ordering for update, where
Tanks and Melees prioritize the close/in spots. This cannot be
implemented as-is just with custom strings.

Add a new strategy option which implements the DN priority bate
strategy. I tried to make a shared function it ended up being worse than
just duplicating the local lambda function used here.

In addition to modifying the alert callout for when to be in/out and
close/far, also modify the combined starting callout. Instead of calling
"close/far" for the baits, call the order of who baits, i.e.:

  "In (Tanks) => Out (Healers) => In (Melee) => Out (Ranged)"

Note that we don't need to call "tanks close/far" because this strategy
always has tanks and melee baiting close while range and healers bait
far.

We could determine if the player is the appropriate target and switch
the order, calling "Party Close (<role> Far)". This changes the trigger
text which might help some folks, but makes it more difficult to parse
overall, in my opinion. In addition, it is not reliably possible to
parse the DPS spots, because not all parties have 2 melee or ranged
players, and often assign someone to be "fake melee".

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>