raidboss: collapse multiple sidewise spark hits to intercards
The way clones and the boss hit for sidewise spark often results in an
intercardinal spot being safe for multiple cleaves. When making a call
for a safe spot, if an intercard is safe for the entire sequence, just
call that intercard.

For the collected callouts where there may be a future callout with a
new safe spot, include the number of hits. This should be two, though
the current code has support for anywhere from 2 to 5.

For the final hit after the boss cleave, just call the direction and
"stay", so its clear the player can just move to that intercard and stay
without needing to keep track of a count.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>