personal: trial r3s fusefield count based on vuln fade
Trigger an info message after every vulnerability fade during fuse
field, informing that the vuln has faded. This isn't strictly perfect,
since players should go only after they are healed as well. This doesn't
use a timer either, so if someone is delayed, this will delay everyone
else. However, players *cannot* disarm while magic vuln is up, so for a
functioning strategy, this should be ok.

In addition, whoever is first should know they need to do the short fuse
immediately, mitigating the risk of a delayed start.

The intent of this trigger is primarily to help a raid caller keep pace
if other players need some reminding when to go.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>