personal: call tower soak for tank based on who got hit
Currently, a single callout of 30 seconds is put up to warn tanks to
soak/swap the tower.

Instead, keep track of which tank got hit, and call out soak or swap
based on who got hit by the tank tower. This enables us to avoid calling
the direction for a tank who just needs to stand in the tower (reducing
noise) and use an alarm vs an alert so tanks can keep track of the beat
of the fight and not forget to swap in to the tower.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>