personal: tweak everkeep forged track callouts
The callout for the forged tracks is somewhat difficult for me to parse.
There is not much time to manage parsing the existing callout before its
too late.

First, prioritize calling the safe direction from wind/fire. The edge
will be safe if you use knockback immunity for the wind. Call it as
"<direction> fire" or "<direction> wind".

Second, reword the inside left and inside right callouts to make it more
clear that they mean the direction if you're looking into the center
from that edge.

Third, add an additional trigger which is timed with a delaySeconds to
callout a knockback alarm when the sword hits a wind elemental line.
This trigger is timed so that its save to use KB immunity once it is
called. Move the final forged track cleanup to this late trigger. For
this reason, it is not marked as a conditional because we need the
cleanup to execute regardless.

These changes should make it easier to process the trigger, and more
reliable to time the kb for the wind.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>