personal: avoid unnecessary wait time for Forged Track
The delay values used for the combination of triggers in the Forged
Track callout is unnecessarily high. A few of the triggers have delays
which are entirely unnecessary, as the triggers do not directly
interact. The final trigger has a delay of 0.6 seconds. Based on my
testing, if the other trigger delays are removed, this can be reduced to
0.2 seconds. This is a half second savings, which give more time to
process and move into position.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>