personal: combine no-cleave and sword collect triggers
There is a separate trigger for calling the 'no cleave' variant of the
Chasm of Vollok. This seems simpler, but it causes a bunch of extra
backround triggers.

In addition, the safe spots for the swords happens about 3 seconds prior
to the left or right half room cleave. Knowing the safe spot early could
be useful for a raid caller.

Refactor to use a single trigger for the sword collects. Make a callout
initially, followed by the callout when we know the direction of the
half room cleave.

While at it, normalize the callout to match the way we call out the half
room cleave.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>