personal: limit provoke and shirk to party only in chaotic
Reduce trigger noise during Chaotic Cloud of Darkness by limiting the
generic provoke and shirk triggers to only trigger for party members.
This is important since there are party-based tankswaps that must occur
quite often during phase 1 of the fight, and this can get very noisy.

This also has the effect of making shirk and provoke trigger for players
outside of the party in field content which is potentially useful.