raidboss: add strategy for Zelenia Ex Escelons' Fall
The Escelons Fall Witch Hunt mechanic in Zelenia Ex has a couple of
distinct strategies. There is the strategy for who baits first, (support
or dps) and then the fixed positions with "DPS start in".

Add a strategy option to allow users to configure a callout that simply
says where to start rather than saying which bait is first.

This doesn't yet handle the more complex EF2 strategies, though it does
work for the "braindead" strategy where you don't change how you resolve
EF2 when dealing with donuts and stacks.