personal: improve readability for bloom 2 trigger
The bloom 2 trigger is hard to read since you have to quickly process
the differences between WSW, WNW, etc.

Replace this with a slightly longer trigger text which is a bit easier
to parse. I tried a couple variations with whether you move diagonally
or not, but didn't come up with anything super great.

First, call out east/west side, then call the 3 spots as "in/out" +
"north/south", where north/south indicate whether you're on the north or
south side of the spots. This is easier to read for me than trying to
parse than the 16-way direction indicators.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>