util: add basic support for CEs to encounter_tools
This is a basic initial attempt at supporting critical encounters and
critical engagements from the encounter tools. It relies on the CE
information we recently started generating from the xivapi, as well as
the DirectorUpdate IDs pulled by hand.

It does appear to work properly at least as far as finding encounters
go. It does at least log the encounter name based off its English
translation.

There is likely more work required to verify that no other encounter
logic breaks. Additionally, we might want to improve the make_timeline
support to use the 0x21 log line as an entry check, since this is what
we do in the timeline files that support critical encounters.

This code was originally based on work from Valarnin, but its been
heavily modified to now pull the CE data from the xivapi. This version
only accepts CEs that match the zone we're in so it should have minimal
chance of breaking other flows. I did end up hard coding the
association of zone to encounter list, which could possibly be improved
somehow.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>