util: generate CE info from xivapi
In preparation for adding proper CE support to the encounter tools,
first start generating the encounter data we need. Unfortunately it is
currently unknown how the game generates the DirectorUpdate IDs used in
the network log messages, but we can at least pull the translated names
of the encounters from xivapi.

Add a new gen_ce_info_and_maps.ts generator script which will generate
several files including:

 * resources/bozja_encounters.ts
 * resources/bozja_ce_map.ts
 * resources/zadnor_encounters.ts
 * resources/zadnor_ce_map.ts
 * resources/south_horn_encounters.ts
 * resources/south_horn_ce_map.ts

I opted to split by zone since that will be useful both to limit
including only the relevent CEs in a given trigger file, and verify that
the zone is correct when we see a relevant network log message in the
encounter tools. This does lead to a lot of extra files. I wasn't sure
how to combine the file results given the way xivapi.writeFile() works.

The list of DirectorUpdate IDs is maintained as part of ce_overrides.ts,
which I tried to structure in a similar way as the zone_overrides.ts.
Hopefully we can re-use that style if we need to override any data from
the API in the future.

I have not made any attempt to support pulling the other translations
for the Chinese and Korean languages yet. I also only pulled the names,
since none of the other data was very relevant for the encounter tools.

Suggested-by: Valarnin <valarnin@users.noreply.github.com>
Signed-off-by: Jacob Keller <jacob.keller@gmail.com>