util: add timeline-offset option for make_timeline
The encounter tools argument parsing as the '-a/--adjust' option which
appears to allow adjusting a timeline file by a given offset. This
argument appears to be mutually exclusive between different modes and
also doesn't even appear to actually be used in anyway.

Remove this argument and replace it with a -o/--timeline-offset argument
of the make_timeline.ts script. This new argument sets the initial
timeline position.

Update the starting header lines to make use of timelinePosition instead
of hardcoding it to 0. This allows users to set the start position which
is especially useful for critical encounters which share many separate
fight timelines within the same timeline file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>