raidboss: improve callouts for Hinkypunk
Update the triggers for Shades Nest and Shades Crossing during
Hinkypunk. Separate the IDs used for the attack from the boss with the
attacks from the clones. For the boss attack in the center, keep the
"in" and "intercards" callouts.

For the attacks from the clones, capture the position data and parse it
to determine the position of the attack, and add this direction to the
callout. This will result in "In at W" or "Intercards from N" for
example to better reflect what is happening.

This is only ~2.5 seconds from when the mechanic triggers, but its
simpler than attempting to predict where the clones are jumping. We
could theoretically check the skill IDs for the telegraph and store the
results for later. However, that is significantly more complex logic,
and I haven't had time to determine the associated skill IDs.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>