personal: raidboss: call Feather rain move 0.3 seconds later
The Feather Rain "move" callout is a timeline trigger that is set to
trigger 3 seconds before the ability hits. This ability doesn't have a
cast or NetworkStartsCasting line. The game expects players to use the
scream from Garuda as a signifier of when to move. The current 3 seconds
is slightly before the scream. While your position *is* snapshot at this
point, most players would not be able to tell precisely when this
snapshot happens. Additionally, if you move *immediately* as the trigger
casts, uncertainty in the timeline could mean you move before the aoe
snapshots your position.

Delay the trigger by reducing the beforeSeconds to 2.7 so that the move
callout is 0.3 seconds later. This will align the callout slightly
better to the scream, and make it less likely to be called out for
moving too "early" by players not using Cactbot. It will also reduce the
risk of mistakenly moving before the snapshot lands if the timeline is
not perfectly synced for any reason.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>