raidboss: WIP: add A15C to ignore abilities for death claw
I experienced a weird issue with timeline breaking on Death Claw
recently, which I have not yet been able to confirm if it was a display
issue in browsingway or something else yet. While testing I discovered
that A15C is obviously an auto attack or other entry that we need to
ignore when generating the timeline. Add a note about this.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>