raidboss: add basic triggers for Arkveld Extreme
Add several basic triggers for various mechanics in the Arkveld Extreme
fight. This includes some callouts for aoes, the chainblade blow, the
limit cut mechanic, and more.

The head markers do appear to be static, at least across multiple logs I
have.

I've tested this in game with a few pulls and things seem to be working
now, but it likely needs more testing.

There are also a few mechanics which I didn't find a good way to make a
trigger and left them out, including the baits and towers, as well as
all of the expanding exaline mechanics.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>